[FIX] Move Section to its own namespace (#52)
* Split content and section

* Change Section namespace

* Make shit runnable

* Fix tests

* Dotnet format

* Remove whitespace

* Fix shit

* Fix merge

* Format

* Equals
