+++ SELearning.API/Controllers/SectionController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Identity.Web.Resource;
 using SELearning.Core.Permission;
+using SELearning.Infrastructure.Authorization;
 
 namespace SELearning.API.Controllers;
 
@@ -12,9 +13,9 @@ namespace SELearning.API.Controllers;
 public class SectionController : ControllerBase
 {
     private readonly ILogger<SectionController> _logger;
-    private readonly IContentService _service;
+    private readonly ISectionService _service;
 
-    public SectionController(ILogger<SectionController> logger, IContentService service)
+    public SectionController(ILogger<SectionController> logger, ISectionService service)
     {
         _logger = logger;
         _service = service;
