+++ SELearning.API/GlobalUsings.cs
@@ -1,9 +1,8 @@
 global using Microsoft.Identity.Web;
 global using System.Security.Claims;
 global using Microsoft.EntityFrameworkCore;
-
 global using SELearning.Core;
-global using SELearning.Core.Content;
 global using SELearning.Core.Comment;
+global using SELearning.Core.Content;
+global using SELearning.Core.Section;
 global using SELearning.Infrastructure;
-global using SELearning.Infrastructure.Authorization;
