+++ SELearning.Core/Content/IContentRepository.cs
@@ -9,12 +9,4 @@ public interface IContentRepository
     public Task<IReadOnlyCollection<ContentDto>> GetContent();
     public Task<Option<ContentDto>> GetContent(int id);
     public Task<IEnumerable<ContentDto>> GetContentByAuthor(string userId);
-
-    // Section
-    public Task<(OperationResult, SectionDto)> AddSection(SectionCreateDto section);
-    public Task<OperationResult> UpdateSection(int id, SectionUpdateDto section);
-    public Task<OperationResult> DeleteSection(int id);
-    public Task<IReadOnlyCollection<SectionDto>> GetSections();
-    public Task<Option<SectionDto>> GetSection(int id);
-    public Task<IReadOnlyCollection<ContentDto>> GetContentInSection(int id);
 }
