+++ SELearning.Core/Content/IContentService.cs
@@ -11,12 +11,4 @@ public interface IContentService
     public Task IncreaseContentRating(int id);
     public Task DecreaseContentRating(int id);
     public Task<IEnumerable<ContentDto>> GetContentByAuthor(string userId);
-
-    // Section
-    public Task<SectionDto> AddSection(SectionCreateDto section);
-    public Task DeleteSection(int id);
-    public Task UpdateSection(int id, SectionUpdateDto section);
-    public Task<IReadOnlyCollection<SectionDto>> GetSections();
-    public Task<SectionDto> GetSection(int id);
-    public Task<IReadOnlyCollection<ContentDto>> GetContentInSection(int id);
 }
