+++ SELearning.Infrastructure.Tests/Content/ContentManagerTests.cs
@@ -1,4 +1,3 @@
-using SELearning.Core.Content;
 using System;
 using System.Threading.Tasks;
 using SELearning.Core.User;
@@ -76,16 +75,6 @@ public class ContentManagerTests : IDisposable
     }
 
     [Fact]
-    public async Task DeleteSection_do_not_throw_error_and_deletes_section()
-    {
-        await _manager.DeleteSection(1);
-
-        var entity = await _context.Section.FindAsync(1);
-
-        Assert.Null(entity);
-    }
-
-    [Fact]
     public async Task DeleteContent_do_not_throw_error_and_deletes_section()
     {
         await _manager.DeleteContent(1);
@@ -95,20 +84,7 @@ public class ContentManagerTests : IDisposable
         Assert.Null(entity);
     }
 
-    [Fact]
-    public async Task CreateSectionAsync_creates_new_content_with_generated_id()
-    {
-        var contentList = new List<Content>();
-        var section = new SectionCreateDto { Title = "title", Description = "description" };
 
-        await _manager.AddSection(section);
-
-        var option = await _repository.GetSection(2);
-
-        Assert.NotNull(option.Value.Id);
-        Assert.Equal("title", option.Value.Title);
-        Assert.Equal("description", option.Value.Description);
-    }
 
     [Fact]
     public async Task CreateContentAsync_creates_new_content_with_generated_id()
@@ -149,26 +125,6 @@ public class ContentManagerTests : IDisposable
     }
 
     [Fact]
-    public async Task ReadSectionAsync_returns_all_Sections()
-    {
-        var allSections = await _manager.GetSections();
-
-        Assert.Collection(allSections,
-            section => Assert.Equal(section.Id, _section.Id)
-        );
-    }
-
-    [Fact]
-    public async Task GetSection_by_id_returns_section()
-    {
-        var section = await _manager.GetSection(1);
-
-        Assert.Equal(1, section.Id);
-        Assert.Equal("python", section.Title);
-        Assert.Equal("description", section.Description);
-    }
-
-    [Fact]
     public async Task GetContent_returns_all_Content()
     {
         var allContent = await _manager.GetContent();
@@ -179,35 +135,6 @@ public class ContentManagerTests : IDisposable
             content => Assert.Equal(3, content.Id),
             content => Assert.Equal(4, content.Id)
         );
-    }
-
-    [Fact]
-    public async Task GetContentInSection_returns_all_Content()
-    {
-        var allContent = await _manager.GetContentInSection(1);
-
-        Assert.Collection(allContent,
-            content => Assert.Equal(1, content.Id),
-            content => Assert.Equal(2, content.Id),
-            content => Assert.Equal(3, content.Id),
-            content => Assert.Equal(4, content.Id)
-        );
-    }
-
-    [Fact]
-    public async Task UpdateSectionAsync_updates_existing_section()
-    {
-        var updateSection = new SectionUpdateDto
-        {
-            Title = "new title",
-            Description = "description",
-        };
-
-        await _manager.UpdateSection(1, updateSection);
-
-        var option = await _repository.GetSection(1);
-
-        Assert.Equal(option.Value.Title, updateSection.Title);
     }
 
     protected virtual void Dispose(bool disposing)
