+++ SELearning.Infrastructure.Tests/Content/ContentRepositoryTests.cs
@@ -51,158 +51,6 @@ public class ContentRepositoryTests
         _repository = new ContentRepository(_context);
     }
 
-    /*
-        Section Tests Below
-    */
-
-    [Fact]
-    public async Task CreateSectionAsync_creates_new_content_with_generated_id()
-    {
-        var section = new SectionCreateDto { Title = "title", Description = "description" };
-
-        var created = (await _repository.AddSection(section)).Item2;
-
-        Assert.NotNull(created.Id);
-        Assert.Equal("title", created.Title);
-        Assert.Equal("description", created.Description);
-    }
-
-    [Fact]
-    public async Task CreateSectionAsync_given_Section_returns_Section_with_Section()
-    {
-        var section = new SectionCreateDto { Title = "title", Description = "description" };
-
-        var (status, created) = await _repository.AddSection(section);
-
-        var sectionDto = new SectionDto { Id = 2, Title = "title", Description = "description" };
-
-        Assert.Equal(sectionDto.Id, created.Id);
-        Assert.Equal(sectionDto.Title, created.Title);
-        Assert.Equal(sectionDto.Description, created.Description);
-        Assert.Equal(OperationResult.Created, status);
-    }
-
-    [Fact]
-    public async Task ReadSectionAsync_given_non_existing_id_returns_None()
-    {
-        var option = await _repository.GetSection(42);
-
-        Assert.True(option.IsNone);
-    }
-
-    [Fact]
-    public async Task UpdateSectionAsync_given_non_existing_id_returns_NotFound()
-    {
-        var section = new SectionUpdateDto
-        {
-            Title = "title",
-            Description = "description",
-        };
-
-        var reponse = await _repository.UpdateSection(42, section);
-
-        Assert.Equal(OperationResult.NotFound, reponse);
-    }
-
-    [Fact]
-    public async Task ReadSectionAsync_returns_all_Sections()
-    {
-        var allSections = await _repository.GetSections();
-
-        Assert.Collection(allSections,
-            section => Assert.Equal(section.Id, _section.Id)
-        );
-    }
-
-    [Fact]
-    public async Task UpdateSectionAsync_updates_existing_section()
-    {
-        var section = new SectionUpdateDto
-        {
-            Title = "title",
-            Description = "description",
-        };
-
-        var updated = await _repository.UpdateSection(1, section);
-
-        Assert.Equal(OperationResult.Updated, updated);
-    }
-
-    [Fact]
-    public async Task UpdateSectionAsync_given_non_existing_Content_returns_NotFound()
-    {
-        var section = new SectionUpdateDto
-        {
-            Title = "title",
-            Description = "description",
-        };
-
-        var response = await _repository.UpdateSection(42, section);
-
-        Assert.Equal(OperationResult.NotFound, response);
-    }
-
-    [Fact]
-    public async Task UpdateSectionAsync_updates_and_returns_Updated()
-    {
-        var contentList = new List<Content>();
-        var section = new SectionUpdateDto
-        {
-            Title = "new title",
-            Description = "description",
-        };
-
-        var response = await _repository.UpdateSection(1, section);
-
-        var entity = await _context.Section.FirstAsync(c => c.Title == "new title");
-
-        Assert.Equal(OperationResult.Updated, response);
-    }
-
-    [Fact]
-    public async Task DeleteSectionAsync_given_non_existing_Id_returns_NotFound()
-    {
-        var response = await _repository.DeleteSection(42);
-
-        Assert.Equal(OperationResult.NotFound, response);
-    }
-
-    [Fact]
-    public async Task DeleteSectionAsync_deletes_and_returns_Deleted()
-    {
-        var response = await _repository.DeleteSection(1);
-
-        var entity = await _context.Section.FindAsync(1);
-
-        Assert.Equal(OperationResult.Deleted, response);
-        Assert.Null(entity);
-    }
-
-    [Fact]
-    public async Task GetContentInSection_returns_Content()
-    {
-        var contentInSection = await _repository.GetContentInSection(1);
-
-        var content = from c in _section.Content
-                      select new ContentDto
-                      {
-                          Id = c.Id,
-                          Author = c.Author,
-                          Title = c.Title,
-                          Description = c.Description,
-                          Section = c.Section,
-                          VideoLink = c.VideoLink,
-                          Rating = c.Rating
-                      };
-
-        Assert.Equal(content, contentInSection);
-    }
-
-
-    /*
-        Contnet Tests Below
-    */
-
     [Fact]
     public async Task CreateContentAsync_creates_new_content_with_generated_id()
     {
