+++ SELearning.Infrastructure.Tests/GlobalUsings.cs
@@ -1,6 +1,6 @@
-﻿global using SELearning.Core;
-global using SELearning.Core.Permission;
-global using SELearning.Core.Content;
+﻿global using Moq;
 global using SELearning.Core.Comment;
-global using Moq;
+global using SELearning.Core.Content;
+global using SELearning.Core.Permission;
+global using SELearning.Core.Section;
 global using System.Collections.Generic;
