+++ SELearning.Infrastructure.Tests/Section/GlobalUsings.cs
@@ -1,0 +1,4 @@
+global using Microsoft.Data.Sqlite;
+global using Microsoft.EntityFrameworkCore;
+global using Xunit;
+global using SELearning.Core;
