+++ SELearning.Infrastructure/GlobalUsings.cs
@@ -1,7 +1,5 @@
 global using Microsoft.EntityFrameworkCore;
-global using System.ComponentModel.DataAnnotations;
-
 global using SELearning.Core;
-global using SELearning.Core.Content;
 global using SELearning.Core.Comment;
+global using SELearning.Core.Content;
 global using SELearning.Core.User;
