+++ SELearning.Infrastructure/SELearningContext.cs
@@ -1,3 +1,4 @@
+using SELearning.Core.Section;
 using SELearning.Core.User;
 
 namespace SELearning.Infrastructure;
