+++ SELearning/Pages/CreateContent.razor
@@ -1,5 +1,7 @@
 @using Microsoft.AspNetCore.Components.WebAssembly.Authentication
 @using SELearning.Core.Content
+@using System.ComponentModel.DataAnnotations
+@using SELearning.Core.Section
 @using System.Net;
 
 @page "/CreateContent"
