+++ SELearning/Pages/CreateSection.razor
@@ -1,3 +1,4 @@
+@using Core.Section
 @using Microsoft.AspNetCore.Components.WebAssembly.Authentication
 @using SELearning.Core.Content
 @using System.Net
